refactor(pattern): split integration tests into detection and configuration
